+++ README.md
@@ -1,5 +1,14 @@
 # human-intent-sdk
 
-人类意图识别SDK
+意图识别SDK
 
+## 主要
+
+识别到多个意图的时候返回二次确认
+
+与平台无关，参数最小化
+
+插件式的组合起来使用
+
+有上下文，至少能延续上一轮对话
