+++ src/main/java/com/bytehonor/sdk/intent/human/recognize/IntentRecognizer.java
@@ -1,17 +1,16 @@
 package com.bytehonor.sdk.intent.human.recognize;
 
-import java.util.Set;
-
 import com.bytehonor.sdk.intent.human.model.IntentRequest;
+import com.bytehonor.sdk.intent.human.model.IntentResult;
 import com.bytehonor.sdk.intent.human.model.IntentTarget;
 
 public interface IntentRecognizer {
-
-    public Set<String> apps();
 
     public String intent();
 
     public String pattern();
 
     public IntentTarget recoginze(IntentRequest request);
+    
+    public IntentResult resolve(IntentTarget target);
 }
