+++ src/main/java/com/bytehonor/sdk/intent/human/util/IntentRecognizeUtils.java
@@ -1,8 +1,5 @@
 package com.bytehonor.sdk.intent.human.util;
 
-import org.springframework.util.CollectionUtils;
-
-import com.bytehonor.sdk.intent.human.constant.IntentConstants;
 import com.bytehonor.sdk.intent.human.recognize.IntentRecognizer;
 import com.bytehonor.sdk.lang.spring.string.SpringString;
 
@@ -12,12 +9,12 @@ public class IntentRecognizeUtils {
         if (SpringString.isEmpty(app) || recognizer == null) {
             return false;
         }
-        if (CollectionUtils.isEmpty(recognizer.apps())) {
-            return false;
-        }
-        if (recognizer.apps().contains(IntentConstants.ANY) == false && recognizer.apps().contains(app) == false) {
-            return false;
-        }
+//        if (CollectionUtils.isEmpty(recognizer.apps())) {
+//            return false;
+//        }
+//        if (recognizer.apps().contains(IntentConstants.ANY) == false && recognizer.apps().contains(app) == false) {
+//            return false;
+//        }
         return true;
     }
 }
